+++ src/com/dean/ordered_collections/tree/ordered_map.clj
@@ -64,14 +64,12 @@
   clojure.lang.MapEquivalence
 
   clojure.lang.Seqable
-  (seq [this]
-    (with-ordered-map this
-      (map node/-kv (tree/node-seq root))))
+  (seq [_]
+    (tree/entry-seq root (tree/node-size root)))
 
   clojure.lang.Reversible
-  (rseq [this]
-    (with-ordered-map this
-      (map node/-kv (tree/node-seq-reverse root))))
+  (rseq [_]
+    (tree/entry-seq-reverse root (tree/node-size root)))
 
   clojure.lang.ILookup
   (valAt [this k not-found]
@@ -192,11 +190,10 @@
     cmp)
   (entryKey [_ entry]
     (key entry))  ;; extract key from MapEntry
-  (seq [this ascending]
-    (with-ordered-map this
-      (if ascending
-        (map node/-kv (tree/node-seq root))
-        (map node/-kv (tree/node-seq-reverse root)))))
+  (seq [_ ascending]
+    (if ascending
+      (tree/entry-seq root)
+      (tree/entry-seq-reverse root)))
   (seqFrom [this k ascending]
     (with-ordered-map this
       (let [[lt present gt] (tree/node-split root k)]
@@ -204,13 +201,13 @@
           ;; ascending: entries with keys >= k
           (if present
             (cons (clojure.lang.MapEntry. (first present) (second present))
-                  (map node/-kv (tree/node-seq gt)))
-            (seq (map node/-kv (tree/node-seq gt))))
+                  (tree/entry-seq gt))
+            (tree/entry-seq gt))
           ;; descending: entries with keys <= k
           (if present
             (cons (clojure.lang.MapEntry. (first present) (second present))
-                  (map node/-kv (tree/node-seq-reverse lt)))
-            (seq (map node/-kv (tree/node-seq-reverse lt))))))))
+                  (tree/entry-seq-reverse lt))
+            (tree/entry-seq-reverse lt))))))
 
   clojure.lang.IHashEq
   (hasheq [this]